El .exe, compilado y ejecutado: tres arreglos del cruce a MinGW
Con el toolchain instalado, el build cruzado dejó de ser teoría. Cuatro
hallazgos, tres arreglados aquí y el cuarto medido.

M_PI no existe en MinGW: matrix.h:23 no compilaba. No es de C++
estándar; glibc la da siempre y MinGW solo con _USE_MATH_DEFINES
definida ANTES de <cmath>, así que va en CPPFLAGS y no en un header.
Inocua fuera de Windows.

Los objetos del cruce no pueden compartir obj/ con los del nativo:
alternar `make` y `make CROSS=…` dejaba objetos ELF que el ld de MinGW
no lee, y el error —«undefined reference to kExtraTextGlyphs»— no dice
nada de eso. El cruce usa obj-win/ y clean borra los dos.

-s al enlazar: 15.6 MB → 1.9 MB. El .exe se descarga, no se depura.

Cinco warnings que clang no da y g++ sí (misleading-indentation ×2,
sign-compare ×3), arreglados para conservar el «compilación limpia» del
Code style ahora que hay dos toolchains. Cero cambio de salida.

MEDICIÓN, la que no se podía hacer de otro modo: la salida de Windows
contra la de Linux, byte a byte, con el .exe bajo wine —24 ejemplos × 3
formatos, normalizando el %%Title del EPS igual que el harness—:

    idénticos 68/72 · difieren 4 · errores 0

Los 24 SVG y los 24 EPS son IDÉNTICOS. Difieren cuatro PDF, los cuatro
por la misma cadena: 0.000000000000000061232 contra
0.00000000000000006123. Es cos(90°) —el cero numérico de
PDFDisplay::deviceRotate— y los cuatro archivos son los que rotan un
rótulo de eje 90°. La diferencia la pone HPDF_FToA de libharu, que saca
los decimales de un log10() y emite con modff, todo en float: hereda la
libm de la plataforma. Detalle y opciones en la bitácora.

Verificado además el único código exclusivo de Windows, exeDir(), con un
.mg de biblioteca que existe SOLO junto al .exe —no en /usr/local, para
que wine no lo resolviera por el camino de Unix y diera un falso
positivo— invocado desde otro directorio de trabajo.

+++ src/text_parser.cpp
@@ -720,6 +720,9 @@ parse_text_core(const string &input, FontFace ff, bool &using_reencode, bool &us
   // asigna a int (→ -1) y luego `v_end == string::npos` vuelve a convertir -1 a
   // size_t máx, así que la comparación funciona. Se deja como está (cambiarlo a
   // size_t requeriría revisar `it = v_end - 1` y `v_end > it` con `it` int).
+  // v_end es int a propósito: recoge el size_t que devuelve find_first_*, y cuando
+  // eso es npos la truncación lo deja en -1, que es contra lo que se compara abajo
+  // —con el cast explícito, si no g++ avisa de comparar con y sin signo—.
   int v_end, it=0, iend = input.size();
   // Modo matemático: '_'/'^' (sub/superíndice) solo operan entre '$…$'; fuera son
   // caracteres literales (así prosa como line_width no se parte en subíndice).
@@ -745,7 +748,7 @@ parse_text_core(const string &input, FontFace ff, bool &using_reencode, bool &us
         break;
       } else if (input[it]=='\\') {
         v_end = input.find_first_not_of(ALFABETIC, ++it);
-        if (v_end == string::npos) // Last word in string
+        if (v_end == (int)string::npos) // Last word in string
           v_end = iend;
         if (v_end > it && v_end <= iend) {
           string variable = input.substr(it, v_end - it);
@@ -762,7 +765,7 @@ parse_text_core(const string &input, FontFace ff, bool &using_reencode, bool &us
         //it = v_end-1;
       } else {
         v_end = input.find_first_of(TOKENLIM, it);
-        if (v_end == string::npos) // Last word in string
+        if (v_end == (int)string::npos) // Last word in string
           v_end = iend;
         if (it >= v_end){
           //printf("punto %d %d %c\n", it, v_end, input[it]);
@@ -880,7 +883,7 @@ parse_text_core(const string &input, FontFace ff, bool &using_reencode, bool &us
       }
       textflush();
       v_end = input.find_first_not_of(ALFABETIC, ++it);
-      if (v_end == string::npos) // Last word in string
+      if (v_end == (int)string::npos) // Last word in string
         v_end = iend;
       if (v_end > it && v_end <= iend) {
         string variable = input.substr(it, v_end - it);
